+++ Modules/Microsoft365DSC/Modules/M365DSCDRGUtil.psm1
@@ -3551,6 +3551,7 @@ function Update-IntuneDeviceConfigurationPolicy
         $TemplateReferenceId,
 
         [Parameter()]
+        [AllowNull()]
         [System.String]
         $CreationSource,
 
@@ -3581,7 +3582,7 @@ function Update-IntuneDeviceConfigurationPolicy
             $policy.Add('templateReference', @{ 'templateId' = $TemplateReferenceId })
         }
 
-        if ($PSBoundParameters.ContainsKey('CreationSource'))
+        if ($PSBoundParameters.ContainsKey('CreationSource') -and -not [System.String]::IsNullOrEmpty($CreationSource))
         {
             $policy.Add('creationSource', $CreationSource)
         }