+++ Tests/Unit/Microsoft365DSC/Microsoft365DSC.IntuneDiskEncryptionFileVaultPolicyMacOS.Tests.ps1
@@ -50,15 +50,221 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
 
             Mock -CommandName Get-MgBetaDeviceManagementConfigurationPolicy -MockWith {
                 return @{
-                    AdditionalProperties = @{
-                        '@odata.type' = "#microsoft.graph.DeviceManagementConfigurationPolicy"
-                    }
-                    Description = "FakeStringValue"
-                    DisableEntraGroupPolicyAssignment = $True
-                    Id = "FakeStringValue"
-                    Name = "FakeStringValue"
+                    Id              = '12345-12345-12345-12345-12345'
+                    Description     = 'My Test'
+                    Name            = 'My Test'
                     RoleScopeTagIds = @("FakeStringValue")
+                    TemplateReference = @{
+                        TemplateId = 'e688156f-6564-4c03-b34f-83b90fe6bb82_1'
+                    }
                 }
+            }
+
+            Mock -CommandName Update-IntuneDeviceConfigurationPolicy -MockWith {
+            }
+
+            Mock -CommandName Get-IntuneSettingCatalogPolicySetting -MockWith {
+            }
+
+            Mock -CommandName Get-MgBetaDeviceManagementConfigurationPolicySetting -MockWith {
+                return @(
+                    @{
+                        Id = '0'
+                        SettingDefinitions = @(
+                            @{
+                                Id = 'com.apple.mcx.filevault2_deferdontaskatuserlogout'
+                                Name = 'DeferDontAskAtUserLogout'
+                                OffsetUri = 'DeferDontAskAtUserLogout'
+                                AdditionalProperties = @{
+                                    '@odata.type' = '#microsoft.graph.deviceManagementConfigurationChoiceSettingDefinition'
+                                    options = @(
+                                        @{
+                                            name = 'Disabled'
+                                            itemId = 'com.apple.mcx.filevault2_deferdontaskatuserlogout_false'
+                                            dependentOn = @(
+                                                @{
+                                                    dependentOn = 'com.apple.mcx.filevault2_com.apple.mcx.filevault2'
+                                                    parentSettingId = 'com.apple.mcx.filevault2_com.apple.mcx.filevault2'
+                                                }
+                                            )
+                                            optionValue = @{
+                                                '@odata.type' = '#microsoft.graph.deviceManagementConfigurationStringSettingValue'
+                                                value = 'false'
+                                            }
+                                        }
+                                        @{
+                                            name = 'Enabled'
+                                            itemId = 'com.apple.mcx.filevault2_deferdontaskatuserlogout_true'
+                                            dependentOn = @(
+                                                @{
+                                                    dependentOn = 'com.apple.mcx.filevault2_com.apple.mcx.filevault2'
+                                                    parentSettingId = 'com.apple.mcx.filevault2_com.apple.mcx.filevault2'
+                                                }
+                                            )
+                                            optionValue = @{
+                                                '@odata.type' = '#microsoft.graph.deviceManagementConfigurationStringSettingValue'
+                                                value = 'true'
+                                            }
+                                        }
+                                    )
+                                }
+                            },
+                            @{
+                                Id = 'com.apple.mcx.filevault2_deferforceatuserloginmaxbypassattempts'
+                                Name = 'DeferForceAtUserLoginMaxBypassAttempts'
+                                OffsetUri = 'DeferForceAtUserLoginMaxBypassAttempts'
+                                AdditionalProperties = @{
+                                    '@odata.type' = '#microsoft.graph.deviceManagementConfigurationSimpleSettingDefinition'
+                                    dependentOn = @(
+                                        @{
+                                            dependentOn = 'com.apple.mcx.filevault2_com.apple.mcx.filevault2'
+                                            parentSettingId = 'com.apple.mcx.filevault2_com.apple.mcx.filevault2'
+                                        }
+                                    )
+                                }
+                            },
+                            @{
+                                Id = 'com.apple.mcx.filevault2_defer'
+                                Name = 'Defer'
+                                OffsetUri = 'Defer'
+                                AdditionalProperties = @{
+                                    '@odata.type' = '#microsoft.graph.deviceManagementConfigurationChoiceSettingDefinition'
+                                    options = @(
+                                        @{
+                                            name = 'Enabled'
+                                            itemId = 'com.apple.mcx.filevault2_defer_true'
+                                            dependentOn = @(
+                                                @{
+                                                    dependentOn = 'com.apple.mcx.filevault2_com.apple.mcx.filevault2'
+                                                    parentSettingId = 'com.apple.mcx.filevault2_com.apple.mcx.filevault2'
+                                                }
+                                            )
+                                            optionValue = @{
+                                                '@odata.type' = '#microsoft.graph.deviceManagementConfigurationStringSettingValue'
+                                                value = 'true'
+                                            }
+                                        }
+                                    )
+                                }
+                            },
+                            @{
+                                Id = 'com.apple.mcx.filevault2_com.apple.mcx.filevault2'
+                                Name = 'com.apple.MCX.FileVault2'
+                                OffsetUri = 'com.apple.MCX.FileVault2'
+                                AdditionalProperties = @{
+                                    '@odata.type' = '#microsoft.graph.deviceManagementConfigurationSettingGroupCollectionDefinition'
+                                    minimumCount = 0
+                                    maximumCount = 1
+                                    childIds = @(
+                                        'com.apple.mcx.filevault2_defer'
+                                        'com.apple.mcx.filevault2_deferdontaskatuserlogout'
+                                        'com.apple.mcx.filevault2_deferforceatuserloginmaxbypassattempts'
+                                    )
+                                }
+                            }
+                        )
+                        SettingInstance = @{
+                            SettingDefinitionId = 'com.apple.mcx.filevault2_com.apple.mcx.filevault2'
+                            SettingInstanceTemplateReference = @{
+                                SettingInstanceTemplateId = '0e20e909-e28a-41a1-8541-5aadd4714d7d'
+                            }
+                            AdditionalProperties = @{
+                                '@odata.type' = '#microsoft.graph.deviceManagementConfigurationGroupSettingCollectionInstance'
+                                groupSettingCollectionValue = @(
+                                    @{
+                                        children = @(
+                                            @{
+                                                '@odata.type' = '#microsoft.graph.deviceManagementConfigurationChoiceSettingInstance'
+                                                settingDefinitionId = 'com.apple.mcx.filevault2_defer'
+                                                choiceSettingValue = @{
+                                                    '@odata.type' = '#microsoft.graph.deviceManagementConfigurationStringSettingValue'
+                                                    children = @()
+                                                    value = 'com.apple.mcx.filevault2_defer_true'
+                                                }
+                                            }
+                                            @{
+                                                '@odata.type' = '#microsoft.graph.deviceManagementConfigurationChoiceSettingInstance'
+                                                settingDefinitionId = 'com.apple.mcx.filevault2_deferdontaskatuserlogout'
+                                                choiceSettingValue = @{
+                                                    '@odata.type' = '#microsoft.graph.deviceManagementConfigurationStringSettingValue'
+                                                    children = @()
+                                                    value = 'com.apple.mcx.filevault2_deferdontaskatuserlogout_false'
+                                                }
+                                            }
+                                            @{
+                                                '@odata.type' = '#microsoft.graph.deviceManagementConfigurationSimpleSettingInstance'
+                                                settingDefinitionId = 'com.apple.mcx.filevault2_deferforceatuserloginmaxbypassattempts'
+                                                simpleSettingValue = @{
+                                                    '@odata.type' = '#microsoft.graph.deviceManagementConfigurationIntegerSettingValue'
+                                                    value = 5
+                                                }
+                                            }
+                                        )
+                                    }
+                                )
+                            }
+                        }
+                    }
+                    @{
+                        Id = '1'
+                        SettingDefinitions = @(
+                            @{
+                                Id = 'com.apple.security.fderecoverykeyescrow_com.apple.security.fderecoverykeyescrow'
+                                Name = 'com.apple.security.FDERecoveryKeyEscrow'
+                                OffsetUri = 'com.apple.security.FDERecoveryKeyEscrow'
+                                AdditionalProperties = @{
+                                    '@odata.type' = '#microsoft.graph.deviceManagementConfigurationSettingGroupCollectionDefinition'
+                                    minimumCount = 0
+                                    maximumCount = 1
+                                    childIds = @(
+                                        'com.apple.security.fderecoverykeyescrow_location'
+                                        'com.apple.security.fderecoverykeyescrow_devicekey'
+                                    )
+                                }
+                            }
+                            @{
+                                Id = 'com.apple.security.fderecoverykeyescrow_location'
+                                Name = 'Location'
+                                OffsetUri = 'Location'
+                                AdditionalProperties = @{
+                                    '@odata.type' = '#microsoft.graph.deviceManagementConfigurationSimpleSettingDefinition'
+                                    dependentOn = @(
+                                        @{
+                                            dependentOn = 'com.apple.security.fderecoverykeyescrow_com.apple.security.fderecoverykeyescrow'
+                                            parentSettingId = 'com.apple.security.fderecoverykeyescrow_com.apple.security.fderecoverykeyescrow'
+                                        }
+                                    )
+                                }
+                            }
+                        )
+                        SettingInstance = @{
+                            SettingDefinitionId = 'com.apple.security.fderecoverykeyescrow_com.apple.security.fderecoverykeyescrow'
+                            SettingInstanceTemplateReference = @{
+                                SettingInstanceTemplateId = '6f9cb7be-2e50-408a-b24b-6e1387a07fc7'
+                            }
+                            AdditionalProperties = @{
+                                '@odata.type' = '#microsoft.graph.deviceManagementConfigurationGroupSettingCollectionInstance'
+                                groupSettingCollectionValue = @(
+                                    @{
+                                        children = @(
+                                            @{
+                                                '@odata.type' = '#microsoft.graph.deviceManagementConfigurationSimpleSettingInstance'
+                                                settingDefinitionId = 'com.apple.security.fderecoverykeyescrow_location'
+                                                simpleSettingValue = @{
+                                                    '@odata.type' = '#microsoft.graph.deviceManagementConfigurationStringSettingValue'
+                                                    value = 'FakeStringValue'
+                                                }
+                                            }
+                                        )
+                                    }
+                                )
+                            }
+                        }
+                    }
+                )
+            }
+
+            Mock -CommandName Update-DeviceConfigurationPolicyAssignment -MockWith {
             }
 
             Mock -CommandName New-M365DSCConnection -MockWith {
@@ -72,18 +278,42 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
             $Script:ExportMode = $false
 
             Mock -CommandName Get-MgBetaDeviceManagementConfigurationPolicyAssignment -MockWith {
+                return @(@{
+                    Id       = '12345-12345-12345-12345-12345'
+                    Source   = 'direct'
+                    SourceId = '12345-12345-12345-12345-12345'
+                    Target   = @{
+                        DeviceAndAppManagementAssignmentFilterId   = '12345-12345-12345-12345-12345'
+                        DeviceAndAppManagementAssignmentFilterType = 'none'
+                        AdditionalProperties                       = @(
+                            @{
+                                '@odata.type' = '#microsoft.graph.exclusionGroupAssignmentTarget'
+                                groupId       = '26d60dd1-fab6-47bf-8656-358194c1a49d'
+                            }
+                        )
+                    }
+                })
             }
-
         }
 
         # Test contexts
         Context -Name "The IntuneDiskEncryptionFileVaultPolicyMacOS should exist but it DOES NOT" -Fixture {
             BeforeAll {
                 $testParams = @{
-                    Description = "FakeStringValue"
-                    DisableEntraGroupPolicyAssignment = $True
-                    Id = "FakeStringValue"
-                    Name = "FakeStringValue"
+                    Assignments = [CimInstance[]]@(
+                        (New-CimInstance -ClassName MSFT_DeviceManagementConfigurationPolicyAssignments -Property @{
+                            DataType = '#microsoft.graph.exclusionGroupAssignmentTarget'
+                            groupId = '26d60dd1-fab6-47bf-8656-358194c1a49d'
+                            deviceAndAppManagementAssignmentFilterType = 'none'
+                        } -ClientOnly)
+                    )
+                    Description = "My Test"
+                    Id = "12345-12345-12345-12345-12345"
+                    DisplayName = "My Test"
+                    Defer                                  = "true";
+                    DeferDontAskAtUserLogout               = "false";
+                    DeferForceAtUserLoginMaxBypassAttempts = 5;
+                    Location                               = "FakeStringValue";
                     RoleScopeTagIds = @("FakeStringValue")
                     Ensure = "Present"
                     Credential = $Credential;
@@ -108,12 +338,22 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
         Context -Name "The IntuneDiskEncryptionFileVaultPolicyMacOS exists but it SHOULD NOT" -Fixture {
             BeforeAll {
                 $testParams = @{
-                    Description = "FakeStringValue"
-                    DisableEntraGroupPolicyAssignment = $True
-                    Id = "FakeStringValue"
-                    Name = "FakeStringValue"
+                    Assignments = [CimInstance[]]@(
+                        (New-CimInstance -ClassName MSFT_DeviceManagementConfigurationPolicyAssignments -Property @{
+                            DataType = '#microsoft.graph.exclusionGroupAssignmentTarget'
+                            groupId = '26d60dd1-fab6-47bf-8656-358194c1a49d'
+                            deviceAndAppManagementAssignmentFilterType = 'none'
+                        } -ClientOnly)
+                    )
+                    Description = "My Test"
+                    Id = "12345-12345-12345-12345-12345"
+                    DisplayName = "My Test"
+                    Defer                                  = "true";
+                    DeferDontAskAtUserLogout               = "false";
+                    DeferForceAtUserLoginMaxBypassAttempts = 5;
+                    Location                               = "FakeStringValue";
                     RoleScopeTagIds = @("FakeStringValue")
-                    Ensure = 'Absent'
+                    Ensure = "Absent"
                     Credential = $Credential;
                 }
             }
@@ -135,12 +375,22 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
         Context -Name "The IntuneDiskEncryptionFileVaultPolicyMacOS Exists and Values are already in the desired state" -Fixture {
             BeforeAll {
                 $testParams = @{
-                    Description = "FakeStringValue"
-                    DisableEntraGroupPolicyAssignment = $True
-                    Id = "FakeStringValue"
-                    Name = "FakeStringValue"
+                    Assignments = [CimInstance[]]@(
+                        (New-CimInstance -ClassName MSFT_DeviceManagementConfigurationPolicyAssignments -Property @{
+                            DataType = '#microsoft.graph.exclusionGroupAssignmentTarget'
+                            groupId = '26d60dd1-fab6-47bf-8656-358194c1a49d'
+                            deviceAndAppManagementAssignmentFilterType = 'none'
+                        } -ClientOnly)
+                    )
+                    Description = "My Test"
+                    Id = "12345-12345-12345-12345-12345"
+                    DisplayName = "My Test"
+                    Defer                                  = "true";
+                    DeferDontAskAtUserLogout               = "false";
+                    DeferForceAtUserLoginMaxBypassAttempts = 5;
+                    Location                               = "FakeStringValue";
                     RoleScopeTagIds = @("FakeStringValue")
-                    Ensure = 'Present'
+                    Ensure = "Present"
                     Credential = $Credential;
                 }
             }
@@ -153,12 +403,22 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
         Context -Name "The IntuneDiskEncryptionFileVaultPolicyMacOS exists and values are NOT in the desired state" -Fixture {
             BeforeAll {
                 $testParams = @{
-                    Description = "FakeStringValue"
-                    DisableEntraGroupPolicyAssignment = $True
-                    Id = "FakeStringValue"
-                    Name = "FakeStringValue"
+                    Assignments = [CimInstance[]]@(
+                        (New-CimInstance -ClassName MSFT_DeviceManagementConfigurationPolicyAssignments -Property @{
+                            DataType = '#microsoft.graph.exclusionGroupAssignmentTarget'
+                            groupId = '26d60dd1-fab6-47bf-8656-358194c1a49d'
+                            deviceAndAppManagementAssignmentFilterType = 'none'
+                        } -ClientOnly)
+                    )
+                    Description = "My Test"
+                    Id = "12345-12345-12345-12345-12345"
+                    DisplayName = "My Test"
+                    Defer                                  = "true";
+                    DeferDontAskAtUserLogout               = "true"; # Drift
+                    DeferForceAtUserLoginMaxBypassAttempts = 5;
+                    Location                               = "FakeStringValue";
                     RoleScopeTagIds = @("FakeStringValue")
-                    Ensure = 'Present'
+                    Ensure = "Present"
                     Credential = $Credential;
                 }
             }
@@ -173,7 +433,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
 
             It 'Should call the Set method' {
                 Set-TargetResource @testParams
-                Should -Invoke -CommandName Update-MgBetaDeviceManagementConfigurationPolicy -Exactly 1
+                Should -Invoke -CommandName Update-IntuneDeviceConfigurationPolicy -Exactly 1
             }
         }